+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Global broadcast texture will now set corresponding scale transform property.
 - Make it compatible with VRCLightVolumes 3.0.0.dev-13 beyond.
 
+### Fixed
+- Luminance don't have effects on VRCLightVolume shaders.
+
 ## 1.7.6-beta.3 - 2026-07-09
 ### Added
 - Color palette picker