+++ CHANGELOG.md
@@ -5,13 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 1.7.6-beta.4 - 2026-08-01
+### Fixed
+- Luminance don't have effects on VRCLightVolume shaders.
+
 ### Changed
 - Auto hide idle screen UI if idle texture been set.
 - Global broadcast texture will now set corresponding scale transform property.
 - Make it compatible with VRCLightVolumes 3.0.0.dev-13 beyond.
-
-### Fixed
-- Luminance don't have effects on VRCLightVolume shaders.
 
 ## 1.7.6-beta.3 - 2026-07-09
 ### Added